% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_base.R
\name{theme_base}
\alias{theme_base}
\title{ColourblindR script for theme_base (Internal function)}
\usage{
theme_base(colour_type)
}
\description{
: theme_base()
}
\details{
This function implements a base theme that makes plots more intepretable for viewers in general
It modifies the layout of the plot.

: Create your plot via ggplot2 and apply our theme to enhance plots to be more readable for colourblind viewers
}
