% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_prota.R
\name{theme_prota}
\alias{theme_prota}
\title{ColourblindR script for protanopia}
\usage{
: ggplot(data, aes(variable x, variable y, colour = class)) + geom_point() + theme_prota('colour')
}
\arguments{
\item{colour_type}{either fill colour of colour depending on the geom of choice}
}
\value{
reformatted plot for protanopia vision
}
\description{
: theme_deutera() creates a theme with a Protanopia friendly colour palette that can be used on
both colour and fill aesthetics.
Use it in conjuction with ggplot2
}
\details{
This function implements a theme that makes plots more intepretable for viewers with protanopia
It modifies the colour of geometric objects (points, lines, etc) and the layout of the plot.

: Create your plot via ggplot2 and apply our theme to enhance plots to be more readable for colourblind viewers
}
\examples{
:

p <- ggplot(mpg, aes(x = hwy, y = cty, fill = as.factor(cyl))) +geom_bar(stat = 'identity')
cowplot::plot_grid(p, p + theme_prota(colour_type = 'fill'))

p <- ggplot(mpg, aes(x = hwy, y = cty, colour = as.factor(cyl))) + geom_point(stat = 'identity')
cowplot::plot_grid(p, p +  theme_prota(colour_type = 'colour'))

ggplot(mpg, aes(hwy, colour = as.factor(cyl))) + geom_density()
cowplot::plot_grid(p, p +  theme_prota(colour_type = 'colour'))

}
