% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_trita.R
\name{theme_trita}
\alias{theme_trita}
\title{ColourblindR script for tritanopia}
\usage{
: ggplot(data, aes(variable x, variable y, colour = class)) + geom_point() + theme_trita('colour')
}
\arguments{
\item{colour_type}{either fill colour of colour depending on the geom of choice}
}
\value{
reformatted plot for tritanopia vision
}
\description{
: theme_deutera() creates a theme with a Tritanopia friendly colour palette that can be used on
both colour and fill aesthetics.
Use it in conjuction with ggplot2
}
\details{
This function implements a theme that makes plots more intepretable for viewers with tritanopia
It modifies the colour of geometric objects (points, lines, etc) and the layout of the plot.

: Create your plot via ggplot2 and apply our theme to enhance plots to be more readable for colourblind viewers
}
\examples{

\dontrun{ggplot(mpg, aes(x = hwy, y = cty, fill = as.factor(cyl))) +geom_bar(stat = 'identity') + theme_trita(colour_type = 'fill')}

\dontrun{ ggplot(mpg, aes(x = hwy, y = cty, colour = as.factor(cyl))) + geom_point(stat = 'identity') +  theme_trita(colour_type = 'colour')}

\dontrun{ggplot(mpg, aes(hwy, colour = as.factor(cyl))) + geom_density()  +  theme_trita(colour_type = 'colour')}

}
